requirements.txt
pytest
pytest-selenium
termcolor
allure-pytest==2.8.18